+++ Assets/UniGLTF/Editor/UniGLTF/ScriptedImporter/TextureExtractor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 556e0ca7f0824e44b9000d04e0e9914c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
